Chore: 임시저장 및 등록하기 버튼 별도 컴포넌트화하여 반응형 구현
